requirements.txt
mediapipe=0.8.4.2
numpy=1.19.5
opencv-contrib-python=4.5.2.52
pillow=8.2.0
scikit-learn=0.24.2
streamlit=0.82.0
tensorflow=2.5.0